requirements.txt
fastapi
passlib
python-jose
python-dotenv
pymongo
mysql-connector-python
python-multipart
uvicorn